mcp_server/instagram/pyproject.toml
[project]
name = "instagram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "instagrapi>=2.1.3",
    "mcp[cli]>=1.3.0",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
]
